pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "youtube-transcript-collector"
version = "0.1.0"
description = "A lightweight Python package for downloading YouTube channel transcripts"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "youtube-transcript-api>=1.2.3",
    "rich>=13.7.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/youtube_transcript_collector"]

[dependency-groups]
dev = ["ipykernel>=7.1.0", "ipywidgets>=8.1.8"]
